Store pnpm stores, caches, and runtime data in the docker image
This means `pnpm` is not required on the host at all (same as with all the python deps), which is pleasing.  It does mean the images must be rebuilt when `package.json` or `pnpm-lock.yaml` are updated, but this isn't new.  It's a tiny bit hacky (but the cleanest I could come up with after much experimentation) so wants keeping half an eye on, but so far looks good, and the advantages are manifold.