Move pose-tracking dependencies to api container
Because ByteTrack should now be runnable as part of the api container, rather than just in the notebook, we need to add scipy to the Pipfile and also install that annoying lap dependency separately. There may be a better way to do this (other than just using something other than ByteTrack).