Handle edge cases of $currentFrame being undefined while switching videos
These sanity checks may not be the ideal fix, but they do stop some errors from appearing in the console after selecting a different video in the video chooser at the top of the interface. This maybe could be in a separate branch, but it's a pretty small change that shouldn't affect anything else.