Highlight frames with close matches to the selected pose in the pose data timelines
Some misuse of the MultiLine chart's display conventions is done to ensure that the lines indicating pose matches use the full height of the chart viewer and thus are easier to see. We may ultimately decide to use a different type of visualization entirely for this, and/or enhance the chart functionality (such as by adding support for multiple Y axes) so that this kind of hack isn't necessary.