Provision a separate Docker container for other web content (specifically PixPlot)
PixPlot instances are fully static once generated, so in theory they could coexist with other static front-end content on the web-ui container. But given that Astro is currently running there in dev mode, for now it's easier just to run PixPlot on a different container, also in its "dev" mode.