Facilitate using a production build of the front end
With these changes the front end will serve a production build from `web-ui/dist` if one exists, and fall back to the dev server if not.  A `just build-prod-ui` command is provided for convenience.  To revert to using the dev server, just `rm -rf` the dist folder.  No container/server restarts or image rebuilds should be required.