Use a materialized view for the video metadata query
View must be created in the `initialize_db` function (or created manually for existing dbs), and refreshed when data is added.  The `just refresh-db-views` command is created for this purpose, and added as a post-dependency to just recipes that touch the relevant tables.  This is not a perfect solution, but will do for now.