Set `<PosesByFrameChart/>` to be an immutable component
There's a bug in Svelte > v3.13.0 that's been open for almost four years that can cause components to render twice in certain circumstances, and we're hitting those here.  This commit is a workaround, setting the component to `immutable` to prevent the issue, and then altering a couple of other pieces so that reactivity still works where we need it.  Apparently the bug will be fixed in the upcoming Svelte 5 (!).