Refactor shenanigans to ensure similar poses/movelets panels stay closed.
After rebasing this branch, some problematic behavior emerged whereby closing the Similar Poses or Similar Movelets panel by clicking the X would close it in the current tab view, but then upon switching to a different tab view, the panel would reappear with the similar poses/movelets displayed from the previous search. Perhaps the change in behavior was due to the new caching optimizations, and so these mitigations probably undermine those performance enhancements? If so, that's a pity, and hopefully we can find a way to get the performance bumps while also providing a (more) consistent UX. But for now, this functionality is preferable, I think.