Reduce the amount of JSON returned per matching pose/movelet
This may no longer be premature optimization, and it does seem to help with load times for the search results. A more aggressive optimization to consider would be to compute the full pose extents and store them in the DB when the pose data is ingested; this would obviate the need to run getExtent() on each set of full pose coordinates, and may make it so we don't need to return the raw pose coordinates from the DB at all (just the normed coordinates). But I'm not so sure about that.