Calculate, store and visualize action "interest"
This is done in exactly the same way as the per-frame pose "interest" is calculated (average deviation from the global mean vector). In fact the pose interest code has by DRYed up to cover the action interest as well. A related change involves storing per-pose interest values in the DB (even if they're not currently visualized). A more incidental change is removing pose counts from the per-frame stats, because in the PHALP output the pose counts are always identical to the track counts for a frame.