Implement frame and pose grabbing
This is a minimal implementation.  The pose still needs to be adjusted when it's drawn on top the captured frame.  There are also some values that are left hard-coded here (mainly because I don't fully understand how they work in this context, and this code will need to be re-generalized at some point in this process), and there may also be some things missing.  But this will do to move on to the next bit.