Draw poses correctly with respect to bboxes
This took rather more false starts that I'd care to admit -- in the end I have done away with previous code and reconsidered from first principles -- I think scaling to fit the canvas is the wrong approach.  This is more complicated than scaling a full frame, since we don't know the aspect ratio ahead of time, but I think this approach is the cleanest.