Unify the source tree for TLS deployment
This is the best way I can come up with to have the necessary config for operation with a self-signed TLS cert be part of the codebase without getting in the way of dev. deployment.  There is duplication in the nginx config files, unfortunately, so it's not a perfect solution, but it's better than what we've got up to now.