Draw the matched hand in the similar hands lineup
in the proper color. Note that there's still a bug wherein the matching
hands are not drawn on the similar hands/poses lineup when the
similarity metric/method is changed in the radio selector. But because
the hands do appear when the pager is advanced or the pose overlays are
hidden and re-displayed, this seems to be an issue of the update not
reaching the Pose elements in that specific case.