Always use PHALP-assigned frame-level pose indices for highlighting
A pose's pose_idx attribute from PHALP can deviate considerably from its position in the pose data array for that frame, due to the presence of untracked poses in the raw output. Mixing these up sometimes causes incorrect behavior when the poses or their FrameDetails entries are hovered. The pose_idx values are consistent and unique to the frame, so we might as well use these exclusively when highlighting poses. When a pose or hand is clicked, however, it's OK to refer to it by position in the data array because in that case the entire pose/hand object is just dumped into the appropriate store to kick off the similarity search.