Invert per-pose and frame avg confidence score
This makes for a somewhat easier to read timeline plot, as the frames with high avg. confidence scores have "badness" (inverted confidence) scores near 0, while frames with lower confidence are highlighted by the spikes in the "BadScore" series. Previously, the "AvgScore" confidence viz was rather confusing and distracting (thus it's hidden by default, which is still true in this branch) because it was usually pegged near its maximum value (1) and thus at the top of the plot, whereas all the other series have 0 as their baseline value. Also, the previous confidence viz paradigm works especially poorly with pose series that have been sampled at rates lower than the video frame rate, as the "avgScore" plot would constantly be oscillating between near 1 for frames with poses and 0 for unsampled frames.