Get avg pose/action interest from the frame table
when displaying them in the frame details pane. Previously, the values displayed apparently were derived arbitrarily from a single pose in the frame. This must have been an oversight. Fortunately the values used to plot the the 'interest' values in the timeline series were always derived from the correct (averaged) values in the frame table.