Fix `video_frame_meta` VIEW creation
Using `1 - pose_details.avg_score` in the SELECT clause results in no `avg_score` column in the view, so queries were throwing `UndefinedColumnError: column "avg_score" does not exist`.  Adding `... AS avg_score` is the obvious fix; however, the `get_post_data_by_frame` endpoint SELECTs `1 - avg_score AS "badScore"`, so the score is inverted again, so reverting the change to this line from 2f7aaad3 is the correct fix.