Remove unused CSS
Both of these selectors and associated rules are removed by the Sveltekit preprocessor, as the selectors don't appear in the component.  There are various ways around this; in the case of the `#main-content` selector that rule is better added to the main `+layout.svelte` file anyway (where it will work because of the `:global()` qualifier), and the `pre` selector seems to be genuinely unused.