Restore frames count to videos table (also in mk1)
and replace "length" with "duration." In principle, the frames count can be calculated from the fps and duration, both of which are already provided in the tables, but who has time for that. I removed the frames count to save space in the table, but since then I've found myself wishing for a convenient way to check that number enough times that it's going back in. It's also worth noting that the frames count could be used to indicate the number of *sampled* frames, when such sampling has been applied, but at present it just shows the number of frames in the original source video.