Fix `cluster_video_faces.py`
When ingesting videos and sidecar files, all files must be placed in the root of the `VIDEO_SRC_FOLDER`; it could be helpful to be able to arrange them in a more organized fashion -- the only reason it's not possible is the way some of the ingest scripts are coded.  I fixed this one and then decided I didn't have time to go through and fix them all -- but might want to come back to this.